graphql filter frage mit id
