+++ app/routes/_index.tsx
@@ -8,103 +8,16 @@ import {useQuery} from "../../node_modules/@apollo/client/react/hooks/useQuery";
 import {gql} from "../../node_modules/graphql-tag/src/index";
 import { Box, Paper, Typography } from '@mui/material';
 import RatingStars from './components/RatingStars';
+import AlleBuecher from '~/graphql/AlleBuecher';
+import BuchMitID from '~/graphql/BuchMitID';
 
 const client = new ApolloClient({
   uri: 'https://localhost:3000/graphql',
   cache: new InMemoryCache(),
 });
 
-console.log("INDEDX")
-
 // TODO: Workaround für Zertifikate finden
 // process.env.NODE_TLS_REJECT_UNAUTHORIZED = '0';
-
-client
-  .query({
-    query: gql`
-      query GetBuecher {
-        buecher {
-          id
-          isbn
-          preis
-          schlagwoerter
-          titel {
-            titel
-          }
-        }
-      }
-    `,
-  })
-  .then((result) => console.log(result.data.buecher));
-
-const GET_BUECHER = gql`
-  query GetBuecher {
-    buecher {
-      id
-      isbn
-      preis
-      schlagwoerter
-      titel {
-        titel
-      }
-    }
-  }
-`;
-
-function DisplayBuecher() {
-  const { loading, error, data } = useQuery(GET_BUECHER);
-  console.log("| LOAD: "+loading+"| ERROR: "+error+"| DATA: "+data)
-  if (loading) return <p>Loading...</p>;
-  if (error) return <p>Error : {error.message}</p>;
-  
-  return (
-    <Box display="flex" flexWrap="wrap" justifyContent="center" gap={2} mt={4}>
-      {data.buecher.map(({ id, isbn, titel, preis,schlagwoerter }) => (
-        <Paper key={id} elevation={3} sx={{ padding: 2, minWidth: 200, textAlign: 'center' }}>
-          <Typography variant="h6">{titel.titel}</Typography>
-          <Typography variant="subtitle1">{schlagwoerter}</Typography>
-          <Typography variant="subtitle1">{isbn}</Typography>
-          <Typography variant="subtitle1">{"UVP: "+ preis+",-"}</Typography>
-        </Paper>
-      ))}
-    </Box>
-  );
-}
-//------------------------------------------
-const FILTER_BOOKS = gql`
-  query BUCH($id: ID!) {
-    buch(id: $id) {
-      id
-      isbn
-      preis
-      schlagwoerter
-      titel {
-        titel
-      }
-    }
-  }
-`;
-
-function BuchMitID({ id }) {
-  const { loading, error, data } = useQuery(FILTER_BOOKS, {
-    variables: { id },
-  });
-
-  if (loading) return null;
-  if (error) return `Error! ${error}`;
-
-  return (
-    <Box display="flex" flexWrap="wrap" justifyContent="center" gap={2} mt={4}>
-        <Paper key={id} elevation={3} sx={{ padding: 2, minWidth: 200, textAlign: 'center' }}>
-          <Typography variant="h6">{data.buch.titel.titel}</Typography>
-          <Typography variant="subtitle1">{data.buch.schlagwoerter}</Typography>
-          <Typography variant="subtitle1">{data.buch.isbn}</Typography>
-          <Typography variant="subtitle1">{"UVP: "+ data.buch.preis+",-"}</Typography>
-        </Paper>
-    </Box>
-  );
-}
-//------------------------------------------
 
 export const meta: MetaFunction = () => {
   return [
@@ -112,6 +25,9 @@ export const meta: MetaFunction = () => {
     { name: "description", content: "Welcome to our Semesterproject!" },
   ];
 };
+
+let sucheBuchID: string | null=null;
+let sucheAlleBuecher=false;
 
 export default function Index() {
   return (
@@ -139,8 +55,8 @@ export default function Index() {
       
       <Box sx={{ flexGrow: 1, padding: '20px' }}>
         <ApolloProvider client={client}>
-          <DisplayBuecher />
-          <BuchMitID id="1"></BuchMitID>
+          <AlleBuecher condition={sucheAlleBuecher}/>
+          <BuchMitID id={sucheBuchID}></BuchMitID>
         </ApolloProvider>
       </Box>
     </Box>
